MakeExcel
I changed the MakeCsv method by the MakeExcel method. I wanted to make an xlsx file but I made a mistake with the format and went for a csv.

Aram's games are retrieved from the start of season 14, i.e. from January 10 to the date on which the command was executed.

to avoid redoing every day I made sure to record the date of the last execution to start again from that day.

I've added a progress bar so you can see how far the method has progressed.

I've also added the fact that if the game has been remade, it's excluded (technically, they're not counted as a win or a loss, normally I think, but if anyone knows the answer, they can make an issue).

I've also added that the excel file should be uploaded to the message at the end.

and for further updates

I think I'll add the champion images (with/instead of) the names

and then we'll see about the rest, like adding graphics, etc.